requirements.txt
instagrapi==2.3.0
anthropic==0.49.0
APScheduler==3.11.0
Pillow==11.1.0
python-dotenv==1.0.1
pytrends==4.9.2
requests==2.32.3
httpx==0.28.1
